This commit signifies the completion of rdt2.2. This protocol handles corrupt packets, but does not handle packet loss. I fully implemented checksum calculations and all state transitions as defined by the rdt2.2 FSM in the textbook.
I decided to implement ACK/NAK as defined in RFC793 (Transmission Control Protocol), which defines the encoding of TCP packets such that a single flag bit is reserved in the TCP header for ACK. This differs from this project's simulation, in which an 8-bit integer is used to represent the ACK flag. Therefore, it might make sense to the student to implement an ACK with the integer value '1' and a NAK with an integer value '-1' (where the Negative Acknolgement is literally the negative value of an Acknolgement). However, I personally decided to limit the value of the ACK integer to either '0' or '1', where '0' means the packet is neither an ACK nor a NAK and '1' means the packet is either a ACK or a NAK. This makes more sense because, in reality, the ACK field is encoded as exactly 1 bit in the header of TCP packets.

In this implementation, whether or not a packet with the ACK/NAK flag enabled (ie: packet->acknum is set to '1') is a (postive) ACK or a (negative) NAK is determined by examining the value of packet->seqnum. In both cases, packet->seqnum is set to the seqnum of the last packet that B successfully recieved from A. Therefore, if the last packet that A sent has a seqnum 0 and the ACK sent back to A from B has a seqnum 0, then A knows that it recieved a (positive) ACK. If the last packet that A sent has a seqnum 0 and the ACK sent back to A from B has a seqnum 1, then A knows that it recieved a (negative) NAK. Similarly, if the last packet that A sent has a seqnum 1 and the ACK sent back to A from B has a seqnum 1, then A knows that it recieved a (positive) ACK. If the last packet that A sent has a seqnum 1 and the ACK sent back to A from B has a seqnum 0, then A knows that it recieved a (negative) NAK.

There are a lot of changes to this commit. I added many new functions and variables. The source code is heavily commented, so I'm not going to attempt to repeat myself. See diff & comments!

TODO
	1) Account for packet loss
		* Implement timeout functionality on both A and B
		* Remove resend on NAK or corrupt (ambigious) ACK; rdt3.0 merely waits for timeout

BUGS
	* For some reason, the simulator doesn't appear to be passing the second packet through the layer3. I'm not sure if this is my fault or the simulators, but the simulator says we shouldn't have to understand how it works or have to modify the code there, so I'll address this issue after rdt3.0 is implemented. In any case, the first packet appears to transmit flawlessly, even with the corruption probability is turned way up (tested at 0.9).